third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/set-breakpoint-while-blocking-main-thread-expected.txt

Tests setting breakpoint when main thread blocks.


Running: testSetBreakpoint
Breakpoint sidebar pane 
blocking-main-thread.js:13checked  Atomics.store(sharedArray, i, 1);
Reloading page.
Page reloaded.
Script execution paused.
Breakpoint sidebar pane 
blocking-main-thread.html:11checked  worker.postMessage(sab);
blocking-main-thread.js:13checked breakpoint hit  Atomics.store(sharedArray, i, 1);
Script execution resumed.
